refactor components -> containing folder
