+++ src/components/overview/ClusterMemory/ClusterMemory.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
-import ClusterTimeseries from '../timeseries/ClusterTimeseries';
+import ClusterTimeseries from '../../timeseries/ClusterTimeseries';
 import uuid from 'uuid';
-import queryLibrary from '../../api/data/queries/query-library';
-import HalinCard from '../ui/scaffold/HalinCard/HalinCard';
+import queryLibrary from '../../../api/data/queries/query-library';
+import HalinCard from '../../ui/scaffold/HalinCard/HalinCard';
 
 class ClusterMemory extends Component {
     state = {
