+++ src/components/overview/GCPauseTime/GCPauseTime.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
-import ClusterTimeseries from '../timeseries/ClusterTimeseries';
+import ClusterTimeseries from '../../timeseries/ClusterTimeseries';
 import uuid from 'uuid';
-import queryLibrary from '../../api/data/queries/query-library';
+import queryLibrary from '../../../api/data/queries/query-library';
 import _ from 'lodash';
-import HalinCard from '../ui/scaffold/HalinCard/HalinCard';
+import HalinCard from '../../ui/scaffold/HalinCard/HalinCard';
 
 class GCPauseTime extends Component {
     state = {
